lambda/requirements.txt
u8darts==0.22.0
boto3==1.25.0
pandas==1.4.4
numpy==1.23.1
python-dotenv==0.21.0